Part 2 complete
All modifications necessary were made to the website, further production testing is needed but overall works. For the local load, only projects show up and the remote has all 5. This is initial to show that the feature works. Part 3 and the ReadMe is still needed to be complete before completion of the homework.